feat(registry): features require a secure context by default, with opt-out (#109)
By default, features will always be "denied" in non-secure contexts. The
"allowed in non-secure contexts" flag undoes this default.

It would be possible to throw an exception from request() and possibly
query() when a secure-only feature is used from a non-secure context,
but denying leads to one fewer behavior that developers need to account
for.