fix(request): clarify how requests affect state
The old wording implied that the user's response to a prompt always
provided "new information" and so would be likely to affect the
permission state, but that interpretation wasn't intended.

The restriction to the same origin also doesn't match what browsers
actually do, especially with respect to ports, but we should fix that in
a separate change.