chore(perm interface): Fix return value handling
Unlike the other methods defined in this section, `revoke` is defined as
a series of steps to be run in parallel. Returning a value is
meaningless in this context. Instead, the result of the operation should
be communicated via the Promise instance that is defined and
synchronously returned prior to the parallelized steps.