Add support for permissions that have an associated Feature Policy
This adds support for denying access to permissions which are not
allowed in the current context based on an associated Feature Policy.
The Feature Name and Default Allowlist of a policy-controlled feature
must still be declared in the owning spec. This also removes specific
support for the media feature policy which is covered by the more
general support.