Add userContext field to WebDriver BiDi's setPermission
User contexts in WebDriver BiDi represents the same concepts
as the user agents in the infra spec: https://w3c.github.io/webdriver-bidi/#user-contexts
According to the infra definition a user agent can be a user profile or a private browsing
window.

This PR adds an ability for the automation client to configure which "user agent"
should the permissions apply too.